// src/Usecase/adCount/types.ts
export type UploadTwitterType = {
    embedded_code: string
    description: string
    time_start: number
    time_expire: number
    category: string
    username: string
    id_iklan: string
}

export type UploadTwitterParam = {
    embedded_code: string
    description: string
    expire: number
}